thou shalt no longer be plagued by vexatious compilation warnings

+++ src/SYSTEM/diag_error.f90
@@ -1,7 +1,7 @@
 subroutine diag_error (info, istyle)
   use iso_c_binding
   implicit none
-  integer(c_long), intent (in) :: info
+  integer, intent (in) :: info
   integer(c_long), intent (in) :: istyle
   write (*,*) '  '
   write (*,*) ' Diagonalization not successful, info = ', info
